requirements.txt
python-telegram-bot
pytumblr
requests
gmusicapi